[M109][DIPS] Add UMA metrics for DIPS Database health
This CL adds metrics for monitoring the health of the SQLite database
which backs DIPSDatabase.

The db's initialization status (success/failure), file size, and entry
count are reported at profile open (which usually happens at startup)
for each non-OTR profile. File size and entry count are also reported
approximately once every 24 hours thereafter, while the profile remains
open and active (triggering db events).

Any SQLite errors that occur while initializing or using the db are
reported at the time of the event.

In addition, this CL also changes the behavior for when DIPSDatabase
fails to open its internal SQLite db. Now, instead of trying to open an
in-memory db if opening one on-disk fails, it tries to open the db
on-disk a second time. If both attempts to open the underlying internal
SQLite db fail, the DIPSDatabase will silently fail, returning
NULL-equivalent values from all methods.

This merge CL also includes the commit for crrev.com/c/3993366, since
it contains dependencies for this commit.

(cherry picked from commit 8e416a7cb702306a999ceaa640ec02eecd241fa6)
(cherry picked from commit 8ebdd2c716cf32da352c3f87121cccc9bc35091d)

Bug: 1394747
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4066763
Commit-Queue: Joshua Hood <jdh@chromium.org>
Cr-Original-Commit-Position: refs/heads/main@{#1078808}
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4080192
Cr-Commit-Position: refs/branch-heads/5414@{#452}
Cr-Branched-From: 4417ee59d7bf6df7a9c9ea28f7722d2ee6203413-refs/heads/main@{#1070088}

+++ chrome/browser/dips/dips_database.cc
@@ -11,6 +11,8 @@
 #include <vector>
 
 #include "base/files/file_path.h"
+#include "base/files/file_util.h"
+#include "base/metrics/histogram_functions.h"
 #include "base/threading/thread_restrictions.h"
 #include "base/time/time.h"
 #include "chrome/browser/dips/dips_utils.h"
@@ -42,6 +44,7 @@ const int kCompatibleVersionNumber = 1;
 // for details.
 
 const base::TimeDelta DIPSDatabase::kMaxAge = base::Days(180);
+const base::TimeDelta DIPSDatabase::kMetricsInterval = base::Hours(24);
 
 DIPSDatabase::DIPSDatabase(const absl::optional<base::FilePath>& db_path)
     : db_path_(db_path.value_or(base::FilePath())),
@@ -53,8 +56,11 @@ DIPSDatabase::DIPSDatabase(const absl::optional<base::FilePath>& db_path)
   if (db_path.has_value()) {
     DCHECK(!db_path->empty())
         << "To create an in-memory DIPSDatabase, explicitly pass an "
-           "absl::nullopt file path.";
+           "absl::nullopt `db_path`.";
   }
+
+  if (Init() != sql::INIT_OK)
+    LOG(WARNING) << "Failed to initialize the DIPS SQLite database.";
 }
 
 DIPSDatabase::~DIPSDatabase() {
@@ -65,8 +71,8 @@ DIPSDatabase::~DIPSDatabase() {
 void DIPSDatabase::DatabaseErrorCallback(int extended_error,
                                          sql::Statement* stmt) {
   DCHECK_CALLED_ON_VALID_SEQUENCE(sequence_checker_);
-  // TODO (jdh@): Try to recover corrupted databases, after we've added the
-  // ability to store the database on disk.
+  sql::UmaHistogramSqliteResult("Privacy.DIPS.DatabaseErrors", extended_error);
+
   if (sql::IsErrorCatastrophic(extended_error)) {
     DCHECK_EQ(1, kCurrentVersionNumber);
 
@@ -121,6 +127,7 @@ bool DIPSDatabase::InitTables() {
 
 sql::InitStatus DIPSDatabase::InitImpl() {
   DCHECK_CALLED_ON_VALID_SEQUENCE(sequence_checker_);
+
   sql::InitStatus status = OpenDatabase();
   if (status != sql::INIT_OK) {
     return status;
@@ -153,13 +160,59 @@ sql::InitStatus DIPSDatabase::Init() {
   DCHECK_CALLED_ON_VALID_SEQUENCE(sequence_checker_);
 
   sql::InitStatus status = InitImpl();
-  if (status == sql::INIT_OK) {
-    return status;
+  int attempts = 1;
+
+  if (status != sql::INIT_OK) {
+    db_->Close();
+
+    // Try to initialize the database once more in case it failed once and was
+    // razed.
+    status = InitImpl();
+    attempts++;
+
+    if (status != sql::INIT_OK) {
+      attempts = 0;
+    }
   }
 
-  db_->Close();
+  base::UmaHistogramExactLinear("Privacy.DIPS.DatabaseInit", attempts, 3);
+
+  last_health_metrics_time_ = base::Time::Now();
+  ComputeDatabaseMetrics();
 
   return status;
+}
+
+void DIPSDatabase::ComputeDatabaseMetrics() const {
+  DCHECK_CALLED_ON_VALID_SEQUENCE(sequence_checker_);
+  base::TimeTicks start_time = base::TimeTicks::Now();
+
+  int64_t db_size;
+  if (base::GetFileSize(db_path_, &db_size)) {
+    base::UmaHistogramMemoryKB("Privacy.DIPS.DatabaseSize", db_size / 1024);
+  }
+
+  base::UmaHistogramCounts10000("Privacy.DIPS.DatabaseEntryCount",
+                                GetEntryCount());
+
+  base::UmaHistogramTimes("Privacy.DIPS.DatabaseHealthMetricsTime",
+                          base::TimeTicks::Now() - start_time);
+}
+
+bool DIPSDatabase::CheckDBInit() const {
+  DCHECK_CALLED_ON_VALID_SEQUENCE(sequence_checker_);
+  if (!db_ || !db_->is_open())
+    return false;
+
+  // Computing these metrics may be costly, so we only do it every
+  // |kMetricsInterval|.
+  base::Time now = base::Time::Now();
+  if (now > last_health_metrics_time_ + kMetricsInterval) {
+    last_health_metrics_time_ = now;
+    ComputeDatabaseMetrics();
+  }
+
+  return true;
 }
 
 bool DIPSDatabase::Write(const std::string& site,
@@ -168,6 +221,9 @@ bool DIPSDatabase::Write(const std::string& site,
                          const TimestampRange& stateful_bounce_times,
                          const TimestampRange& stateless_bounce_times) {
   DCHECK_CALLED_ON_VALID_SEQUENCE(sequence_checker_);
+  if (!CheckDBInit())
+    return false;
+
   static constexpr char kWriteSql[] =  // clang-format off
       "INSERT OR REPLACE INTO bounces("
           "site,"
@@ -198,6 +254,9 @@ bool DIPSDatabase::Write(const std::string& site,
 
 absl::optional<StateValue> DIPSDatabase::Read(const std::string& site) {
   DCHECK_CALLED_ON_VALID_SEQUENCE(sequence_checker_);
+  if (!CheckDBInit())
+    return absl::nullopt;
+
   static constexpr char kReadSql[] =  // clang-format off
       "SELECT site,"
           "first_site_storage_time,"
@@ -231,6 +290,9 @@ absl::optional<StateValue> DIPSDatabase::Read(const std::string& site) {
 
 std::vector<std::string> DIPSDatabase::GetAllSitesForTesting() const {
   DCHECK_CALLED_ON_VALID_SEQUENCE(sequence_checker_);
+  if (!CheckDBInit())
+    return {};
+
   static constexpr char kReadSql[] =  // clang-format off
       "SELECT site FROM bounces ORDER BY site";
   // clang-format on
@@ -249,6 +311,9 @@ std::vector<std::string> DIPSDatabase::GetSitesThatBounced(
     base::Time range_start,
     base::Time last_interaction) const {
   DCHECK_CALLED_ON_VALID_SEQUENCE(sequence_checker_);
+  if (!CheckDBInit())
+    return {};
+
   DCHECK(last_interaction < range_start);
   static constexpr char kReadSql[] =  // clang-format off
       "SELECT site FROM bounces "
@@ -277,6 +342,9 @@ std::vector<std::string> DIPSDatabase::GetSitesThatBouncedWithState(
     base::Time range_start,
     base::Time last_interaction) const {
   DCHECK_CALLED_ON_VALID_SEQUENCE(sequence_checker_);
+  if (!CheckDBInit())
+    return {};
+
   DCHECK(last_interaction < range_start);
   static constexpr char kReadSql[] =  // clang-format off
       "SELECT site FROM bounces "
@@ -304,6 +372,9 @@ std::vector<std::string> DIPSDatabase::GetSitesThatUsedStorage(
     base::Time range_start,
     base::Time last_interaction) const {
   DCHECK_CALLED_ON_VALID_SEQUENCE(sequence_checker_);
+  if (!CheckDBInit())
+    return {};
+
   DCHECK(last_interaction < range_start);
   static constexpr char kReadSql[] =  // clang-format off
       "SELECT site FROM bounces "
@@ -329,6 +400,9 @@ std::vector<std::string> DIPSDatabase::GetSitesThatUsedStorage(
 
 bool DIPSDatabase::RemoveRow(const std::string& site) {
   DCHECK_CALLED_ON_VALID_SEQUENCE(sequence_checker_);
+  if (!CheckDBInit())
+    return false;
+
   static constexpr char kRemoveSql[] = "DELETE FROM bounces WHERE site=?";
   DCHECK(db_->IsSQLValid(kRemoveSql));
 
@@ -342,7 +416,8 @@ bool DIPSDatabase::RemoveEventsByTime(const base::Time& delete_begin,
                                       const base::Time& delete_end,
                                       const DIPSEventRemovalType type) {
   DCHECK_CALLED_ON_VALID_SEQUENCE(sequence_checker_);
-  DCHECK(db_->is_open());
+  if (!CheckDBInit())
+    return false;
 
   sql::Transaction transaction(db_.get());
   if (!transaction.Begin())
@@ -363,6 +438,8 @@ bool DIPSDatabase::ClearTimestamps(const base::Time& delete_begin,
                                    const base::Time& delete_end,
                                    const DIPSEventRemovalType type) {
   DCHECK_CALLED_ON_VALID_SEQUENCE(sequence_checker_);
+  if (!CheckDBInit())
+    return false;
 
   if (type == DIPSEventRemovalType::kAll) {
     static constexpr char kAllTypesSql[] =  // clang-format off
@@ -518,6 +595,8 @@ bool DIPSDatabase::AdjustFirstTimestamps(const base::Time& delete_begin,
                                          const base::Time& delete_end,
                                          const DIPSEventRemovalType type) {
   DCHECK_CALLED_ON_VALID_SEQUENCE(sequence_checker_);
+  if (!CheckDBInit())
+    return false;
 
   if ((type & DIPSEventRemovalType::kHistory) ==
       DIPSEventRemovalType::kHistory) {
@@ -592,6 +671,8 @@ bool DIPSDatabase::AdjustLastTimestamps(const base::Time& delete_begin,
                                         const base::Time& delete_end,
                                         const DIPSEventRemovalType type) {
   DCHECK_CALLED_ON_VALID_SEQUENCE(sequence_checker_);
+  if (!CheckDBInit())
+    return false;
 
   if ((type & DIPSEventRemovalType::kHistory) ==
       DIPSEventRemovalType::kHistory) {
@@ -664,6 +745,8 @@ bool DIPSDatabase::AdjustLastTimestamps(const base::Time& delete_begin,
 
 size_t DIPSDatabase::GetEntryCount() const {
   DCHECK_CALLED_ON_VALID_SEQUENCE(sequence_checker_);
+  if (!CheckDBInit())
+    return 0;
 
   sql::Statement s_entry_count(
       db_->GetCachedStatement(SQL_FROM_HERE, "SELECT COUNT(*) FROM bounces"));
@@ -672,6 +755,8 @@ size_t DIPSDatabase::GetEntryCount() const {
 
 size_t DIPSDatabase::GarbageCollect() {
   DCHECK_CALLED_ON_VALID_SEQUENCE(sequence_checker_);
+  if (!CheckDBInit())
+    return 0;
 
   size_t num_entries = GetEntryCount();
   size_t num_deleted = 0;
@@ -694,6 +779,8 @@ size_t DIPSDatabase::GarbageCollect() {
 
 size_t DIPSDatabase::GarbageCollectExpired() {
   DCHECK_CALLED_ON_VALID_SEQUENCE(sequence_checker_);
+  if (!CheckDBInit())
+    return 0;
 
   base::Time safe_date(base::Time::Now() - kMaxAge);
 
@@ -715,6 +802,8 @@ size_t DIPSDatabase::GarbageCollectExpired() {
 
 size_t DIPSDatabase::GarbageCollectOldest(int purge_goal) {
   DCHECK_CALLED_ON_VALID_SEQUENCE(sequence_checker_);
+  if (!CheckDBInit())
+    return 0;
 
   static constexpr char kGarbageCollectOldestSql[] =  // clang-format off
       "DELETE FROM bounces WHERE site "
